requirements.txt
flask
flask-sqlalchemy
google-generativeai
python-dotenv